feat(66) : 마이페이지 내 입찰 내역 (#76)
* fix : AuctionStatus 수정으로 인한 InitProcess Value 수정

* feat : SecurityConfig에 Swagger 추가

* feat : 마이페이지 내 입찰 구현

+++ src/main/java/com/deal4u/fourplease/domain/member/mypage/dto/HighestBidInfo.java
@@ -1,7 +1,9 @@
 package com.deal4u.fourplease.domain.member.mypage.dto;
 
+import java.math.BigDecimal;
+
 public record HighestBidInfo(
         Long auctionId,
-        Double highestPrice
+        BigDecimal highestPrice
 ) {
 }
